refactor: change ship namespace to ships
