+++ Assets/Scripts/Ships/AIShip.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2a1eda6d5f4c1594c86a1d39cfc2d3fa
