+++ Assets/Scripts/Ships/Modules/Command.cs
@@ -1,4 +1,4 @@
-namespace Ship.Modules
+namespace Ships.Modules
 {
     public class Command : Module
     {
