+++ Assets/Scripts/Ships/Modules/Engine.cs
@@ -1,4 +1,4 @@
-namespace Ship.Modules
+namespace Ships.Modules
 {
     public class Engine : Module
     {
