+++ Assets/Scripts/Ships/Modules/Laser.cs.meta
