+++ Assets/Scripts/Ships/Modules/Module.cs.meta
