+++ Assets/Scripts/Ships/ShipSerializer.cs.meta
